% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restimizeapi.R
\name{About}
\alias{About}
\title{Function prints some information about this package.}
\usage{
About()
}
\description{
Function prints some information about this package.
}
\examples{
\dontrun{
     About()
 }
}

