% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restimizeapi.R
\name{GetCompanies}
\alias{GetCompanies}
\title{Returns the company information for all companies that Estimize covers.}
\usage{
GetCompanies()
}
\value{
The companies information as a data frame.
}
\description{
Returns the company information for all companies that Estimize covers.
}
\examples{
\dontrun{
 result <- GetCompanies ()
 }
}

