% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restimizeapi.R
\name{GetCompany}
\alias{GetCompany}
\title{Returns the company information for the company specified by the ticker.}
\usage{
GetCompany(ticker)
}
\arguments{
\item{ticker}{The Estimize ticker -- for example "MSFT".}
}
\value{
The company information as a data frame.
}
\description{
Returns the company information for the company specified by the ticker.
}
\examples{
\dontrun{
     result <- GetCompany ("msft")
 }
}

