% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restimizeapi.R
\name{GetCompanyEstimates}
\alias{GetCompanyEstimates}
\title{Returns all estimates for a company specified by the ticker.}
\usage{
GetCompanyEstimates(ticker, year = NULL, quarter = NULL)
}
\arguments{
\item{ticker}{The Estimize ticker -- for example "MSFT".}

\item{year}{A four-digit year -- for example 1999.}

\item{quarter}{A numeric value between 1 and 4 inclusive. Note that if the
 quarter is used the year must also be set as well.}
}
\value{
The company estimates as a data frame.
}
\description{
Returns all estimates for a company specified by the ticker.
}
\examples{
\dontrun{
 result <- GetCompanyEstimates ("MSFT")
 result <- GetCompanyEstimates ("MSFT", 2015)
 result <- GetCompanyEstimates ("MSFT", 2015, 2)
 }
}

