% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restimizeapi.R
\name{GetEstimates}
\alias{GetEstimates}
\title{Returns all estimates in the specified date-range for all companies.}
\usage{
GetEstimates(startDate, endDate)
}
\arguments{
\item{startDate}{The start date, which must be in the format YYYY-MM-DD, for
example "2015-01-20"}

\item{endDate}{The end date, which must be in the format YYYY-MM-DD, for
 example "2015-02-15"}
}
\value{
The estimates as a data frame.
}
\description{
Returns all estimates in the specified date-range for all companies.
}
\examples{
\dontrun{
     result <- GetEstimates ("2015-01-20", "2015-02-15")
 }
}

