% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restimizeapi.R
\docType{package}
\name{restimizeapi}
\alias{restimizeapi}
\alias{restimizeapi-package}
\title{Functions for Working with the 'www.estimize.com' Web Services}
\description{
Provides the user with functions to develop their trading
strategy, uncover actionable trading ideas, and monitor consensus shifts with
crowdsourced earnings and economic estimate data directly from
<www.estimize.com>. Further information regarding the web services this
package invokes can be found at <www.estimize.com/api>.
}
\details{


Provides a complete implementation of all web methods available at
at \href{http://api.estimize.com}{api.estimize.com} -- functions available
include:
    \itemize{
    \item \code{\link{GetCompanies}}
    \item \code{\link{GetCompany}}
    \item \code{\link{GetCompanyEstimates}}
    \item \code{\link{GetCompanyReleases}}
    \item \code{\link{GetEstimates}}
    \item \code{\link{GetReleaseConsensus}}
}

Note that the \code{\link{SetKey}} function must be called \emph{once} prior
to invoking any of the other functions in this package.
}

