% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/restimizeapi.R
\name{SetKey}
\alias{SetKey}
\title{Function sets the key which is required as a header in every web method call
to estimize.com -- contact \email{sales@estimize.com} to obtain a key.}
\usage{
SetKey(key)
}
\arguments{
\item{key}{The API key as provided by Estimize.com.}
}
\description{
Function sets the key which is required as a header in every web method call
to estimize.com -- contact \email{sales@estimize.com} to obtain a key.
}
\examples{
\dontrun{
 SetKey ("ENTER YOUR PRIVATE API KEY HERE.")
 }
}

